Cli: added "candidates"  switch and concurrency
This adds the ability for each try to run several different threads in parallel, one for each "candidate". This should make generation faster overall, although it also incurs a cost for duplicating images over and over.

The number of candidates defaults to the number of logical processors in the machine. In case of 1 candidate, concurrency is not used; a different, optimial execution path kicks off. Stats displayed are slightly different as well.

This approach is good, but might be revisited for performance.