Generator: new paint method using the GPU
This is a big change that removes the old method of painting using CPU instructions on an image, and replaces it with a WGPU-drin GPGPU approach.

The big change is that operations now take place over textures, as compute shaders, on a newly created GPU context.

This has many side effects:

- The "Rects" painter has been recreated using shaders
  - It employs a slightly different painting algorithm, and supports antialiasing
- The "Circles" painter has been recreated using shaders
  - The new painting algorithm fixes some errors in the previous implementation
- The "Strokes" painter is temporarily removed (it will be replaced by something new later)
- Blending is done after painting, and with a shader as well
- Calculating diffs is more closely part of the painting process, and also done with a shader
- We don't use multi-threading anymore
- Benchmarks are calculated in a slightly different fashion

This is a major breaking change, as most of the images created previously won't be recreated in the same way, even if the same command line arguments are passed.